knowledge-service/requirements.txt
youtube-transcript-api==0.6.3
youtube-search-python==1.6.6
chromadb==0.5.23
sentence-transformers==3.4.1
flask==3.1.0
flask-cors==5.0.1
anthropic==0.42.0
python-dotenv==1.0.1
tiktoken==0.8.0
requests==2.32.3
beautifulsoup4==4.12.3
PyPDF2==3.0.1
